requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
channels==4.2.0
channels-redis==4.2.1
psycopg[binary]==3.2.4
